minor refactor to distinguish OpenGL files
