+++ engine/source/platform/platformGL.h
@@ -23,7 +23,7 @@
 #ifndef _PLATFORMGL_H_
 // redirect to appropriate platform file
 #if defined(TORQUE_OS_OSX)
-#include "platformOSX/platformGL.h"
+#include "platformOSX/GL/platformGL.h"
 #elif defined(TORQUE_OS_WIN)
 #include "platformWin32/platformGL.h"
 #elif defined(TORQUE_OS_LINUX) || defined(TORQUE_OS_OPENBSD) || defined(TORQUE_OS_FREEBSD)
